feat: move treemap/sunburst/boxplot/trellis charts into atlas-ui
Promotes TreemapChart, SunburstChart, BoxPlotChart, and TrellisChart
into src/components/ui/charts/ with the Atlas prefix, fixes chart-config
imports, repoints 8 call-site importers, adds Histoire stories and
exports all four from packages/atlas-ui/index.ts.
