+++ packages/atlas-ui/index.ts
@@ -12,7 +12,11 @@ export { default as AtlasBanner } from '../../src/components/ui/AtlasBanner.vue'
 export { default as AtlasButton } from '../../src/components/ui/AtlasButton.vue'
 export { default as AtlasCard } from '../../src/components/ui/AtlasCard.vue'
 export { default as AtlasBarChart } from '../../src/components/ui/charts/AtlasBarChart.vue'
+export { default as AtlasBoxPlotChart } from '../../src/components/ui/charts/AtlasBoxPlotChart.vue'
 export { default as AtlasChartExport } from '../../src/components/ui/charts/AtlasChartExport.vue'
+export { default as AtlasSunburstChart } from '../../src/components/ui/charts/AtlasSunburstChart.vue'
+export { default as AtlasTrellisChart } from '../../src/components/ui/charts/AtlasTrellisChart.vue'
+export { default as AtlasTreemapChart } from '../../src/components/ui/charts/AtlasTreemapChart.vue'
 export { default as AtlasCheckbox } from '../../src/components/ui/AtlasCheckbox.vue'
 export { default as AtlasChip } from '../../src/components/ui/AtlasChip.vue'
 export { default as AtlasCol } from '../../src/components/ui/AtlasCol.vue'
