+++ src/components/datasources/ObservationPeriodReport.vue
@@ -161,7 +161,7 @@ import type { ObservationPeriodReport } from '@/models/datasource.types'
 import ChartSection from './shared/ChartSection.vue'
 import EmptyReportState from './shared/EmptyReportState.vue'
 import BarChart from '@/components/reports/charts/BarChart.vue'
-import BoxPlotChart from '@/components/reports/charts/BoxPlotChart.vue'
+import BoxPlotChart from '@/components/ui/charts/AtlasBoxPlotChart.vue'
 import PieChart from '@/components/reports/charts/PieChart.vue'
 import MultiLineChart from './charts/MultiLineChart.vue'
 import DashboardAgeChart from './charts/DashboardAgeChart.vue'
